Remove layout and update graphql queries
