+++ src/pages/404.js
@@ -1,12 +1,15 @@
 import React from 'react';
 
+import GlobalLayout from '../components/Layout/GlobalLayout';
 import Main from '../components/Layout/Main';
 
 const NotFoundPage = () => (
-  <Main>
-    <h1>NOT FOUND</h1>
-    <p>You just hit a route that doesn&#39;t exist&hellip;the sadness.</p>
-  </Main>
+  <GlobalLayout>
+    <Main>
+      <h1>NOT FOUND</h1>
+      <p>You just hit a route that doesn&#39;t exist&hellip;the sadness.</p>
+    </Main>
+  </GlobalLayout>
 );
 
 export default NotFoundPage;
