+++ src/pages/about.js
@@ -1,23 +1,30 @@
 import React from 'react';
+import { StaticQuery, graphql } from 'gatsby';
 
+import GlobalLayout from '../components/Layout/GlobalLayout';
 import Main from '../components/Layout/Main';
 import ContactForm from '../components/Content/ContactForm';
 
-const AboutPage = ({ data }) => (
-  <Main>
-    <h1>About {data.site.siteMetadata.title}</h1>
-    <ContactForm />
-  </Main>
-);
-
-export const query = graphql`
-  query AboutQuery {
-    site {
-      siteMetadata {
-        title
+const AboutPage = () => (
+  <StaticQuery
+    query={graphql`
+      query AboutQuery {
+        site {
+          siteMetadata {
+            title
+          }
+        }
       }
-    }
-  }
-`;
+    `}
+    render={data => (
+      <GlobalLayout>
+        <Main>
+          <h1>About {data.site.siteMetadata.title}</h1>
+          <ContactForm />
+        </Main>
+      </GlobalLayout>
+    )}
+  />
+);
 
 export default AboutPage;
