+++ src/pages/search.js
@@ -1,21 +1,28 @@
 import React from 'react';
+import { StaticQuery, graphql } from 'gatsby';
 
+import GlobalLayout from '../components/Layout/GlobalLayout';
 import Main from '../components/Layout/Main';
 
 const SearchPage = ({ data }) => (
-  <Main>
-    <h1>Search {data.site.siteMetadata.title}</h1>
-  </Main>
-);
-
-export const query = graphql`
-  query SearchQuery {
-    site {
-      siteMetadata {
-        title
+  <StaticQuery
+    query={graphql`
+        query SearchQuery {
+        site {
+          siteMetadata {
+            title
+          }
+        }
       }
-    }
-  }
-`;
+    `}
+    render={data => (
+      <GlobalLayout>
+        <Main>
+          <h1>Search {data.site.siteMetadata.title}</h1>
+        </Main>
+      </GlobalLayout>
+    )}
+  />
+);
 
 export default SearchPage;
